Refactor integration utils into separate files and re-write url_safe_join
No-Issue
